# add fastpath for float images with nans