--- a/pyqtgraph/functions_numba.py
+++ b/pyqtgraph/functions_numba.py
@@ -22,19 +22,26 @@ def rescaleData(data, scale, offset, dtype, clip):
     return data_out
 
 @numba.jit(nopython=True)
-def _rescale_and_lookup1d_function(data, scale, offset, lut, out):
+def rescale_and_lookup(data, scale, offset, lut):
+    # data should be floating point and 2d
+    # lut is 1d
     vmin, vmax = 0, lut.shape[0] - 1
+    out = np.empty_like(data, dtype=lut.dtype)
     for (x, y) in np.nditer((data, out)):
         val = (x - offset) * scale
         val = min(max(val, vmin), vmax)
         y[...] = lut[int(val)]
+    return out
 
-def rescale_and_lookup1d(data, scale, offset, lut):
-    # data should be floating point and 2d
-    # lut is 1d
-    data_out = np.empty_like(data, dtype=lut.dtype)
-    _rescale_and_lookup1d_function(data, float(scale), float(offset), lut, data_out)
-    return data_out
+@numba.jit(nopython=True)
+def rescale_and_clip(data, scale, offset, vmin, vmax):
+    # vmin and vmax <= 255
+    out = np.empty_like(data, dtype=np.uint8)
+    for (x, y) in np.nditer((data, out)):
+        val = (x - offset) * scale
+        val = min(max(val, vmin), vmax)
+        y[...] = val
+    return out
 
 @numba.jit(nopython=True)
 def numba_take(lut, data):