This PR augments #1693 by adding support for float images containing nans. 

To test it out.
```
set PYQTGRAPHPROFILE="ImageItem.paint"
```
Then run the following script and change the levels using the `HistogramLUTItem`. The GUI should feel more responsive with this PR versus master.
```python
import pyqtgraph as pg
import numpy as np
pg.setConfigOptions(imageAxisOrder='row-major')
# pg.setConfigOptions(useNumba=True)

Y = np.sinc(np.linspace(-6, 10, 4000))[:, np.newaxis]
X = np.sinc(np.linspace(-15, 9, 6000))
data = np.abs(Y * X)
data[data > 0.95] = -1
data = 10*np.log10(data)
data = data.astype(np.float32)

pg.mkQApp()
imv = pg.ImageView()
imv.show()
imv.setImage(data, levels=(-50, 0))
hlut = imv.getHistogramWidget()
hlut.gradient.setColorMap(pg.colormap.get("CET-C1"))
hlut.gradient.showTicks(False)
pg.exec()
```

timings from `PYQTGRAPHPROFILE`
|   |row-major master | row-major this PR | col-major master | col-major this PR |
| -------- | -------- | ------- | ------- | ------- |
| numba | 224 ms | 35 ms | 558 ms | 165 ms |
| numpy | 258 ms | 128 ms | 303 ms | 252 ms |

numba codepath with col-major is no longer worse than numpy. If you were interested in performance, you wouldn't be using col-major anyway...