@pijyoi 2024-03-31
While improving the numba performance for the float with nans case, we also found out how to improve the numba performance for the float with no nans case. (24ms --> 12ms)

The gap between numpy vs numba has widened.