@pijyoi 2024-04-02
Change uint16 mono images to use `functions_numba.rescale_only` instead of `functions_numba.rescaleData`.
Timings have halved.