@pijyoi 2024-04-06
Added support for float RGB images with nans.
The supported data types can thus be summarized as (uint8, uint16, float32, float64) x (1 chan, 3 chans) provided no multi-channel levels are used.

```python
import pyqtgraph as pg
import numpy as np
pg.setConfigOptions(imageAxisOrder='row-major')
# pg.setConfigOptions(useNumba=True)

bias = 50

Y = np.sinc(np.linspace(-6, 10, 4000))[:, np.newaxis]
X = np.sinc(np.linspace(-15, 9, 6000))
data = np.abs(Y * X)
mask1 = data > 0.95
mask2 = (0.5 < data) & (data < 0.6)
data = 10*np.log10(data)
data += bias
data = np.dstack((data, data * 0.5, data * 0.25))
data[..., 1][mask1] = np.nan
data[..., 2][mask2] = np.nan
data = data.astype(np.float32)

pg.mkQApp()
imv = pg.ImageView(levelMode='mono')
imv.show()
imv.setImage(data, levels=(-50 + bias, 0 + bias))
pg.exec()
print(imv.imageItem.qimage.format())
```