@j9ac9k 2024-04-17
that numba `rescale_and_clip` function seems to be exactly what's needed haha, wow, that's _really_ fast.