@j9ac9k 2024-04-17
Every time I think you've gotten as much out of ImageItem performance as we can get, you prove me wrong 😆  Thanks for the PR @pijyoi Merging.